--- libraries/rtprocessing/helpers/filter.cpp
@@ -1,6 +1,0 @@
-#include "filter.h"
-
-Filter::Filter()
-{
-
-}
